make all p2p modules observables
